requirements.txt
uvicorn[standard]
fastapi
redis

chromadb
tiktoken
openai
mcp
oss2
playwright
DrissionPage

aiofiles
aiohttp
aiosqlite
httpx
requests

python-dotenv
PyJWT
pydantic
pydantic[email]
pydantic-settings
pandas
openpyxl
PyYAML
python-xlib
PyMuPDF
python-docx

numpy
opencv-python